decapsulate: remove header length guards
Remove the checks for the length of the header. For some of the protocols,
the chosen value was arbitrary. The guard checks also duplicate code:
the length is checked by the appropriate codec for the protocol.

Consequently, remove the final function clause to catch truncated
headers. decapsulate/1,2 should crash instead since the packet is
invalid.