802.1q: store next header's type
To process the frame as a stream of bytes, put the 802.1q TPID into the
ethernet header and store the next header type as part of the 802.1q
protocol header.

The 802.1q header is inserted into the original frame between the source
MAC address and the ethernet type. Multiple 802.1q headers may be
inserted into the frame:

    original frame:
        |dst MAC|src MAC|ethernet type|

    with 802.1q header:
        ethernet header type = 802.1q

        |dst MAC|src MAC|16#8100 (802.1q TPID)|Prio, CFI, VLAN ID|ethernet type|

    with 802.1qinq header:
        ethernet header type = 802.1q QinQ

        |dst MAC|src MAC
        |16#88a8 (802.1q QinQ TPID)|Prio, CFI, VLAN ID
        |16#8100 (802.1q TPID)|Prio, CFI, VLAN ID
        |ethernet type|

Thanks to Alexander Shavelev (@alexshavelev) for reporting!